requirements.txt
certifi==2019.9.11
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==3.2.0
django-crispy-forms==1.7.2
django-debug-toolbar==2.1
django-rest-framework==0.1.0
django-tinymce4-lite==1.7.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
gunicorn==20.0.0
jsmin==2.2.2
psycopg2==2.8.4
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2019.3
qrcode==6.1
six==1.13.0
sqlparse==0.3.0
whitenoise==4.1.4
nltk==3.4.5
joblib==0.13.2